Refactor getFootnotes selector
Code continues working as usual but it has been optimized, made simpler and the
dependency to getFlattenedBlocks has been removed, so it no longer has to be
stored in the utils.
